Add config option to switch between FFTW_MEASURE and FFTW_ESTIMATE
The compile-time option FFTW_PLAN_MEASURE was added in 67e41b6705e7, when it was noticed that fftw locks up for minutes at a time in emscripten when using FFTW_MEASURE because browsers smear (intentionally decrease the precision of) timers that fftw used for measurements, thereby confusing fftw into a near-infinite loop. Using FFTW_ESTIMATE specifically on emscripten solved this.

This commit makes this a runtime option with a compile-time default, because other systems may also smear timers this way, which can cause the same issue to manifest. E.g. we got such a report from a kali (debian) aarch64 system with fftw 3.3.10. Somehow the official aarch64 linux build was unaffected, so this may be a new fftw bug that doesn't yet affect the tpt-libs version 3.3.8.